Terraria 1.4 support added
New Area scripting functionality added for the new biomes
- Graveyard (InGraveYard)
- Gemcave (InGemCave)
- Granite (InGranite)
- Golem Temple (InLihzhardTemple)

Known Issues with this version

- Broken Custom Invasions
- Custom Monster loading is buggy
- Housing spams errors with database issues

More testing needed...

+++ Banking/Rewards/TileKey.cs
@@ -34,7 +34,7 @@ namespace Banking.Rewards
 		/// Terraria Wall Id.
 		/// </summary>
 		[JsonProperty(Order = 1)]
-		public byte Wall { get; set; } = 0;
+		public ushort Wall { get; set; } = 0;
 
 		public TileKey()
 		{
