+++ Corruption/TEdit/Terraria/Tile.cs
@@ -62,7 +62,7 @@ namespace Corruption.TEdit //TEditXNA.Terraria
 		public bool WireYellow;
 		public byte TileColor;
 		public ushort Type;
-		public byte Wall;
+		public ushort Wall;
 		public byte WallColor;
 		public LiquidType LiquidType;
 		public byte LiquidAmount;
